+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- BREAKING: Decouple results store and artefacts store in `run` interface. Changes `run` arguments.
+
+
 ## [v0.1.0] - 2021-09-09
 ### Added
 - Registry interfaces